# packages/shared-types/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JurisdictionProfile": {
        "description": "The public jurisdiction profile. Indexed, no login. Section 10.4.",
        "properties": {
          "approval_rate_by_use_class": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Approval Rate By Use Class",
            "type": "object"
          },
          "bodies": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Bodies",
            "type": "array"
          },
          "decisions": {
            "title": "Decisions",
            "type": "integer"
          },
          "instruments": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Instruments",
            "type": "array"
          },
          "next_elections": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Next Elections",
            "type": "array"
          },
          "summary": {
            "$ref": "#/components/schemas/JurisdictionSummary"
          }
        },
        "required": [
          "summary",
          "approval_rate_by_use_class",
          "decisions",
          "instruments",
          "bodies",
          "next_elections"
        ],
        "title": "JurisdictionProfile",
        "type": "object"
      },
...
      "JurisdictionSummary": {
        "properties": {
          "bodies": {
            "title": "Bodies",
            "type": "integer"
          },
          "civic_platform": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Civic Platform"
          },
          "data_depth": {
            "title": "Data Depth",
            "type": "integer"
          },
          "discretion_index": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discretion Index"
          },
          "elections_known": {
            "title": "Elections Known",
            "type": "integer"
          },
          "freshness": {
            "enum": [
              "fresh",
              "stale",
              "broken",
              "never"
            ],
            "title": "Freshness",
            "type": "string"
          },
          "has_boundary": {
            "title": "Has Boundary",
            "type": "boolean"
          },
          "hours_since_refresh": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hours Since Refresh"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "legal_framework": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Legal Framework"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "name",
          "kind",
          "region",
          "legal_framework",
          "civic_platform",
          "data_depth",
          "discretion_index",
          "bodies",
          "elections_known",
          "has_boundary",
          "freshness",
          "hours_since_refresh"
        ],
        "title": "JurisdictionSummary",
        "type": "object"
      },
      "LocateLink": {
        "description": "One link in the chain of bodies that can say no to a parcel.\n\nNamed for its endpoint rather than `JurisdictionLink`, which already exists in\n`auspice.score.models`. Two schemas with one name make FastAPI disambiguate both by full module path,\nwhich renames the existing one in the OpenAPI document and breaks every consumer that referenced it.",
        "properties": {
          "data_depth": {
            "title": "Data Depth",
            "type": "integer"
          },
          "discretion_index": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discretion Index"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "legal_framework": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Legal Framework"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "name",
          "kind",
          "role",
          "region",
          "legal_framework",
          "data_depth",
          "discretion_index"
        ],
        "title": "LocateLink",
        "type": "object"
      },
      "LocateResponse": {
        "description": "Who decides for a coordinate. The stage 0 question, answered on its own.\n\nSeparate from scoring on purpose. \"Which bodies can say no to this parcel\" is useful before anyone\nasks for a probability, it is the honest answer for a site outside the covered counties, and it is\ncheap: a spatial join rather than a model fit.\n\n``covered`` is false rather than the response being a 404, because a point in an uncovered county is a\nvalid question with a real answer. Returning nothing found would read as a malfunction.",
        "properties": {
          "chain": {
            "items": {
              "$ref": "#/components/schemas/LocateLink"
            },
            "title": "Chain",
            "type": "array"
          },
          "covered": {
            "title": "Covered",
            "type": "boolean"
          },
          "latitude": {
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "title": "Longitude",
            "type": "number"
          },
          "note": {
            "title": "Note",
            "type": "string"
          }
        },
        "required": [
          "longitude",
          "latitude",
          "covered",
          "chain",
          "note"
        ],
        "title": "LocateResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Calibrated permission risk forecasts with a published accuracy record.\n\nEvery response carries an interval and a data date. When the evidence is too thin the response abstains, and an abstention is a successful response rather than an error. The accuracy record and the jurisdiction profiles are public and need no key.",
    "summary": "A rating bureau for the right to build.",
    "title": "Permission Bureau",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/public/jurisdictions/{slug}": {
      "get": {
        "operationId": "jurisdiction_profile_v1_public_jurisdictions__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JurisdictionProfile"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "A jurisdiction profile",
        "tags": [
          "public"
        ]
      }
    },
    "/v1/public/ledger": {
      "get": {
        "description": "The whole ledger, so anyone can verify it without using our interface.\n\nA record that can only be checked through the publisher's own tooling is not a public record, so this\ndefaults to everything and there is no page size the caller cannot turn off. What changed is how it is\nsent: streamed a line at a time rather than assembled into one string first. The record grows forever\nby design, and an endpoint that materialises all of it to serve it is an endpoint whose memory cost\ngrows with the thing the company is trying to accumulate.\n\n``after`` and ``limit`` exist for a consumer fetching in slices, not for us deciding how much of our\nown record to show. Both are optional and neither is applied unless asked for.",
        "operationId": "ledger_export_v1_public_ledger_get",
        "parameters": [
          {
            "description": "Return entries after this sequence number",
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Return entries after this sequence number",
              "minimum": 0,
              "title": "After",
              "type": "integer"
            }
          },
          {
            "description": "Maximum entries to return",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100000,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum entries to return",
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Ledger Export V1 Public Ledger Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "The full prediction ledger, newline delimited JSON",
        "tags": [
          "public"
        ]
      }
    },
    "/v1/public/locate": {
      "get": {
        "description": "Stage 0 on its own: the chain of bodies that can refuse a project at this point.\n\nPublic and unauthenticated like the rest of this router, and safe to be: it reads the boundary index\nthat `auspice registry load` built from published county shapefiles, and returns nothing a visitor\ncould not get from the counties themselves. It is a spatial join with no model fit behind it.\n\nA point outside the covered counties returns ``covered: false`` with an empty chain rather than a 404,\nbecause \"we do not cover this place\" is a real answer to a reasonable question and a 404 reads as a\nbroken endpoint.",
        "operationId": "locate_v1_public_locate_get",
        "parameters": [
          {
            "in": "query",
            "name": "longitude",
            "required": true,
            "schema": {
              "maximum": 180,
              "minimum": -180,
              "title": "Longitude",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "latitude",
            "required": true,
            "schema": {
              "maximum": 90,
              "minimum": -90,
              "title": "Latitude",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Who decides for this coordinate",
        "tags": [
          "public"
        ]
      }
    },
...
  }
}